delete display file
